+++ Kernel/Arch/x86_64/VGA/TextModeConsole.h
